+++ internal/domain/order.go
@@ -4,9 +4,9 @@ import "gorm.io/gorm"
 
 type Order struct {
 	gorm.Model
-	ProductID  uint  `gorm:"not null"`
-	UserID     uint  `gorm:"not null"`
-	PriceCents int64 `gorm:"not null"`
+	ProductID  uint `gorm:"not null"`
+	UserID     uint `gorm:"not null"`
+	PriceCents uint `gorm:"not null"`
 
 	Product Product `gorm:"foreignKey:ProductID"`
 	User    User    `gorm:"foreignKey:UserID"`
